Create py_bank
I had difficulty with the loops and presenting the revenue. I wasn't able to obtain the revenue, so no arithmetic operations could be conducted to have rest of the answers that were requested. I'll be doing more studying on my own with other resources and making sure I understand this as soon as possible.